+++ driver/internal/protocol/auth/jwt.go
@@ -1,9 +1,11 @@
 package auth
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/SAP/go-hdb/driver/internal/protocol/encoding"
+	"github.com/SAP/go-hdb/driver/internal/trace"
 	"golang.org/x/text/transform"
 )
 
@@ -17,7 +19,10 @@ type JWT struct {
 // NewJWT creates a new authJWT instance.
 func NewJWT(token string) *JWT { return &JWT{token: token} }
 
-func (a *JWT) String() string { return fmt.Sprintf("method type %s token %s", a.Typ(), a.token) }
+func (a *JWT) String() string {
+	return fmt.Sprintf("method type %s logonname %s token %s cookie %s",
+		a.Typ(), trace.Cut(a.logonname), trace.Redacted(a.token), trace.Redacted(a._cookie))
+}
 
 // Cookie implements the AuthCookieGetter interface.
 func (a *JWT) Cookie() (string, []byte) { return a.logonname, a._cookie }
@@ -28,29 +33,46 @@ func (a *JWT) Typ() string { return MtJWT }
 // Order implements the Method interface.
 func (a *JWT) Order() byte { return MoJWT }
 
-// PrepareInitReq implements the Method interface.
-func (a *JWT) PrepareInitReq(prms *Prms) error {
-	prms.addString(a.Typ())
-	prms.addString(a.token)
+// AuthLoginName implements the Method interface.
+func (a *JWT) AuthLoginName() string { return a.logonname }
+
+// EncodeInitReq implements the Method interface.
+func (a *JWT) EncodeInitReq(prms *Prms) error {
+	prms.AddString(a.token)
 	return nil
 }
 
-// InitRepDecode implements the Method interface.
-func (a *JWT) InitRepDecode(dec *encoding.Decoder) error {
+// DecodeInitReq implements the Method interface.
+func (a *JWT) DecodeInitReq(dec *encoding.Decoder) error {
+	_, token := dec.LIBytes()
+	a.token = string(token)
+	return nil
+}
+
+// DecodeInitReply implements the Method interface.
+func (a *JWT) DecodeInitReply(dec *encoding.Decoder) error {
 	a.logonname = dec.AuthString()
 	return nil
 }
 
-// PrepareFinalReq implements the Method interface.
-func (a *JWT) PrepareFinalReq(prms *Prms) error {
-	prms.AddCESU8String(a.logonname)
-	prms.addString(a.Typ())
+// EncodeFinalReq implements the Method interface.
+func (a *JWT) EncodeFinalReq(prms *Prms) error {
 	prms.addEmpty() // empty parameter
 	return nil
 }
 
-// FinalRepDecode implements the Method interface.
-func (a *JWT) FinalRepDecode(dec *encoding.Decoder, _ transform.Transformer) error {
+// DecodeFinalReq implements the Method interface.
+func (a *JWT) DecodeFinalReq(dec *encoding.Decoder, logonname string) error {
+	a.logonname = logonname
+	_, empty := dec.LIBytes()
+	if len(empty) != 0 {
+		return errors.New("expected empty parameter")
+	}
+	return nil
+}
+
+// DecodeFinalReply implements the Method interface.
+func (a *JWT) DecodeFinalReply(dec *encoding.Decoder, _ transform.Transformer) error {
 	if err := DecodeAndCheckNumPrm(dec, 2); err != nil {
 		return err
 	}