+++ driver/internal/protocol/auth/sessioncookie.go
@@ -1,9 +1,11 @@
 package auth
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/SAP/go-hdb/driver/internal/protocol/encoding"
+	"github.com/SAP/go-hdb/driver/internal/trace"
 	"golang.org/x/text/transform"
 )
 
@@ -20,7 +22,8 @@ func NewSessionCookie(cookie []byte, logonname, clientID string) *SessionCookie 
 }
 
 func (a *SessionCookie) String() string {
-	return fmt.Sprintf("method type %s cookie %v", a.Typ(), a.cookie)
+	return fmt.Sprintf("method type %s cookie %s logonname %s clientID %s",
+		a.Typ(), trace.Redacted(a.cookie), trace.Cut(a.logonname), trace.Cut(a.clientID))
 }
 
 // Typ implements the Method interface.
@@ -29,28 +32,47 @@ func (a *SessionCookie) Typ() string { return MtSessionCookie }
 // Order implements the Method interface.
 func (a *SessionCookie) Order() byte { return MoSessionCookie }
 
-// PrepareInitReq implements the Method interface.
-func (a *SessionCookie) PrepareInitReq(prms *Prms) error {
-	prms.addString(a.Typ())
-	prms.addBytes(append(a.cookie, a.clientID...)) // cookie + clientID !!!
+// AuthLoginName implements the Method interface.
+func (a *SessionCookie) AuthLoginName() string { return a.logonname }
+
+// EncodeInitReq implements the Method interface.
+func (a *SessionCookie) EncodeInitReq(prms *Prms) error {
+	b := make([]byte, 0, len(a.cookie)+len(a.clientID))
+	b = append(b, a.cookie...)
+	b = append(b, a.clientID...)
+	prms.addBytes(b) // cookie + clientID
 	return nil
 }
 
-// InitRepDecode implements the Method interface.
-func (a *SessionCookie) InitRepDecode(_ *encoding.Decoder) error {
+// DecodeInitReq implements the Method interface.
+func (a *SessionCookie) DecodeInitReq(dec *encoding.Decoder) error {
+	_, a.cookie = dec.LIBytes() // cookie + clientID
 	return nil
 }
 
-// PrepareFinalReq implements the Method interface.
-func (a *SessionCookie) PrepareFinalReq(prms *Prms) error {
-	prms.AddCESU8String(a.logonname)
-	prms.addString(a.Typ())
+// DecodeInitReply implements the Method interface.
+func (a *SessionCookie) DecodeInitReply(_ *encoding.Decoder) error {
+	return nil
+}
+
+// EncodeFinalReq implements the Method interface.
+func (a *SessionCookie) EncodeFinalReq(prms *Prms) error {
 	prms.addEmpty() // empty parameter
 	return nil
 }
 
-// FinalRepDecode implements the Method interface.
-func (a *SessionCookie) FinalRepDecode(dec *encoding.Decoder, _ transform.Transformer) error {
+// DecodeFinalReq implements the Method interface.
+func (a *SessionCookie) DecodeFinalReq(dec *encoding.Decoder, logonname string) error {
+	a.logonname = logonname
+	_, empty := dec.LIBytes()
+	if len(empty) != 0 {
+		return errors.New("expected empty parameter")
+	}
+	return nil
+}
+
+// DecodeFinalReply implements the Method interface.
+func (a *SessionCookie) DecodeFinalReply(dec *encoding.Decoder, _ transform.Transformer) error {
 	if err := DecodeAndCheckNumPrm(dec, 2); err != nil {
 		return err
 	}